Move the Ion value types to the types module
