+++ src/binary/non_blocking/binary_buffer.rs
@@ -7,8 +7,8 @@ use crate::binary::uint::DecodedUInt;
 use crate::binary::var_int::VarInt;
 use crate::binary::var_uint::VarUInt;
 use crate::result::{decoding_error, incomplete_data_error, incomplete_data_error_raw};
-use crate::types::integer::UInt;
-use crate::{Int, IonResult, IonType};
+use crate::types::UInt;
+use crate::{types::Int, IonResult, IonType};
 use num_bigint::{BigInt, BigUint, Sign};
 use std::io::Read;
 use std::mem;
