+++ src/element/element_stream_reader.rs
@@ -1,8 +1,8 @@
 use crate::result::{decoding_error, illegal_operation, illegal_operation_raw};
 use crate::text::parent_container::ParentContainer;
 
-use crate::element::iterators::SymbolsIterator;
 use crate::element::{Blob, Clob, Element};
+use crate::types::iterators::SymbolsIterator;
 use crate::{
     Decimal, Int, IonError, IonReader, IonResult, IonType, Str, StreamItem, Symbol, Timestamp,
 };
@@ -361,8 +361,8 @@ mod reader_tests {
     use super::*;
     use crate::result::IonResult;
     use crate::stream_reader::IonReader;
-    use crate::types::decimal::Decimal;
-    use crate::types::timestamp::Timestamp;
+    use crate::types::Decimal;
+    use crate::types::Timestamp;
 
     use crate::IonType;
