+++ src/element/mod.rs
@@ -25,26 +25,19 @@ use std::fmt::{Display, Formatter};
 
 mod annotations;
 pub mod builders;
-mod bytes;
 mod element_stream_reader;
-mod iterators;
-mod list;
-mod lob;
 pub mod reader;
-mod sequence;
-mod sexp;
-mod r#struct;
 pub mod writer;
 
 // Re-export the Value variant types and traits so they can be accessed directly from this module.
-pub use self::bytes::Bytes;
+pub use crate::types::Bytes;
+pub use crate::types::{Blob, Clob};
 pub use annotations::{Annotations, IntoAnnotations};
-pub use lob::{Blob, Clob};
 
-pub use list::List;
-pub use r#struct::Struct;
-pub use sequence::Sequence;
-pub use sexp::SExp;
+pub use crate::types::List;
+pub use crate::types::SExp;
+pub use crate::types::Sequence;
+pub use crate::types::Struct;
 
 impl IonEq for Value {
     fn ion_eq(&self, other: &Self) -> bool {
@@ -614,7 +607,7 @@ where
 #[cfg(test)]
 mod tests {
     use crate::element::annotations::IntoAnnotations;
-    use crate::types::timestamp::Timestamp;
+    use crate::types::Timestamp;
     use crate::{ion_list, ion_sexp, ion_struct, Decimal, Int, IonType, Symbol};
     use chrono::*;
     use rstest::*;
@@ -864,7 +857,7 @@ mod tests {
     }
 
     use crate::element::{Annotations, Element, IntoAnnotatedElement, Struct};
-    use crate::types::integer::IntAccess;
+    use crate::types::IntAccess;
     use num_bigint::BigInt;
     use std::collections::HashSet;
     use std::str::FromStr;
