+++ src/raw_reader.rs
@@ -1,8 +1,8 @@
 use crate::element::{Blob, Clob};
 use crate::raw_symbol_token::RawSymbolToken;
 use crate::stream_reader::IonReader;
-use crate::types::string::Str;
 use crate::types::IonType;
+use crate::types::Str;
 use crate::{Decimal, Int, IonResult, Timestamp};
 use std::fmt::{Display, Formatter};
 use std::io::Read;
