+++ src/text/parsers/timestamp.rs
@@ -14,8 +14,8 @@ use crate::text::parse_result::{
 };
 use crate::text::parsers::{stop_character, trim_zeros_and_parse_i32, trim_zeros_and_parse_u32};
 use crate::text::text_value::TextValue;
-use crate::types::decimal::Decimal;
-use crate::types::timestamp::{FractionalSecondSetter, Timestamp};
+use crate::types::Decimal;
+use crate::types::{FractionalSecondSetter, Timestamp};
 
 /// Matches the text representation of a timestamp value and returns the resulting Timestamp
 /// as a [TextValue::Timestamp].
@@ -282,8 +282,8 @@ mod reader_tests {
     use crate::text::parsers::timestamp::parse_timestamp;
     use crate::text::parsers::unit_test_support::{parse_test_err, parse_test_ok};
     use crate::text::text_value::TextValue;
-    use crate::types::decimal::Decimal;
-    use crate::types::timestamp::Timestamp;
+    use crate::types::Decimal;
+    use crate::types::Timestamp;
 
     fn parse_equals(text: &str, expected: Timestamp) {
         parse_test_ok(parse_timestamp, text, TextValue::Timestamp(expected))
