+++ src/text/text_writer.rs
@@ -2,8 +2,8 @@ use crate::element::writer::TextKind;
 use crate::raw_symbol_token_ref::{AsRawSymbolTokenRef, RawSymbolTokenRef};
 use crate::result::IonResult;
 use crate::text::raw_text_writer::RawTextWriter;
-use crate::types::decimal::Decimal;
-use crate::types::timestamp::Timestamp;
+use crate::types::Decimal;
+use crate::types::Timestamp;
 use crate::writer::IonWriter;
 use crate::{Int, IonType, RawTextWriterBuilder, SymbolTable};
 use delegate::delegate;
