+++ src/types/decimal.rs
@@ -6,8 +6,8 @@ use num_bigint::{BigInt, BigUint, ToBigUint};
 use crate::ion_data::IonEq;
 use crate::ion_data::IonOrd;
 use crate::result::{illegal_operation, IonError};
-use crate::types::coefficient::{Coefficient, Sign};
-use crate::types::integer::UInt;
+use crate::types::UInt;
+use crate::types::{Coefficient, Sign};
 use num_traits::Zero;
 use std::convert::{TryFrom, TryInto};
 use std::fmt::{Display, Formatter};
@@ -387,8 +387,8 @@ impl TryFrom<Decimal> for BigDecimal {
 #[cfg(test)]
 mod decimal_tests {
     use crate::result::IonResult;
-    use crate::types::coefficient::{Coefficient, Sign};
-    use crate::types::decimal::Decimal;
+    use crate::types::Decimal;
+    use crate::types::{Coefficient, Sign};
     use bigdecimal::BigDecimal;
     use num_bigint::BigUint;
     use num_traits::{Float, ToPrimitive};
