+++ src/types/sexp.rs
@@ -1,34 +1,34 @@
 use crate::element::builders::SequenceBuilder;
-use crate::element::iterators::ElementsIterator;
 use crate::element::{Element, Sequence};
 use crate::ion_data::IonEq;
 use crate::text::text_formatter::IonValueFormatter;
+use crate::types::iterators::ElementsIterator;
 use delegate::delegate;
 use std::fmt::{Display, Formatter};
 
-/// An in-memory representation of an Ion list.
+/// An in-memory representation of an Ion s-expression
 /// ```
 /// use ion_rs::element::{Element, List};
-/// use ion_rs::ion_list;
+/// use ion_rs::ion_sexp;
 /// # use ion_rs::IonResult;
 /// # fn main() -> IonResult<()> {
-/// let list = ion_list![1, 2, 3];
-/// assert_eq!(list.len(), 3);
-/// assert_eq!(list.get(1), Some(&Element::integer(2)));
+/// let sexp = ion_sexp!(1 2 3);
+/// assert_eq!(sexp.len(), 3);
+/// assert_eq!(sexp.get(1), Some(&Element::integer(2)));
 /// # Ok(())
 /// # }
 /// ```
-/// To build a `List` incrementally, see [SequenceBuilder].
+/// To build a `SExp` incrementally, see [SequenceBuilder].
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct List(pub Sequence);
+pub struct SExp(pub Sequence);
 
-impl List {
+impl SExp {
     pub(crate) fn new(elements: impl Into<Sequence>) -> Self {
-        List(elements.into())
+        SExp(elements.into())
     }
 }
 
-impl List {
+impl SExp {
     delegate! {
         to self.0 {
             pub fn clone_builder(&self) -> SequenceBuilder;
@@ -40,21 +40,21 @@ impl List {
     }
 }
 
-impl IonEq for List {
+impl IonEq for SExp {
     fn ion_eq(&self, other: &Self) -> bool {
         // The inner `Sequence` of both Lists are IonEq
         self.0.ion_eq(&other.0)
     }
 }
 
-impl AsRef<Sequence> for List {
+impl AsRef<Sequence> for SExp {
     fn as_ref(&self) -> &Sequence {
         &self.0
     }
 }
 
-// Allows `for element in &list {...}` syntax
-impl<'a> IntoIterator for &'a List {
+// Allows `for element in &sexp {...}` syntax
+impl<'a> IntoIterator for &'a SExp {
     type Item = &'a Element;
     type IntoIter = ElementsIterator<'a>;
 
@@ -63,37 +63,37 @@ impl<'a> IntoIterator for &'a List {
     }
 }
 
-impl Display for List {
+impl Display for SExp {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         let mut ivf = IonValueFormatter { output: f };
-        ivf.format_list(self).map_err(|_| std::fmt::Error)?;
+        ivf.format_sexp(self).map_err(|_| std::fmt::Error)?;
         Ok(())
     }
 }
 
-impl From<Sequence> for List {
+impl From<Sequence> for SExp {
     fn from(sequence: Sequence) -> Self {
-        List(sequence)
+        SExp(sequence)
     }
 }
 
-impl From<List> for Sequence {
-    fn from(value: List) -> Self {
+impl From<SExp> for Sequence {
+    fn from(value: SExp) -> Self {
         value.0
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::ion_list;
-    use crate::types::integer::IntAccess;
+    use crate::ion_sexp;
+    use crate::types::IntAccess;
 
     #[test]
-    fn for_element_in_list() {
-        // Simple example to exercise List's implementation of IntoIterator
-        let list = ion_list![1, 2, 3];
+    fn for_element_in_sexp() {
+        // Simple example to exercise SExp's implementation of IntoIterator
+        let sexp = ion_sexp!(1 2 3);
         let mut sum = 0;
-        for element in &list {
+        for element in &sexp {
             sum += element.as_i64().unwrap();
         }
         assert_eq!(sum, 6i64);
