+++ src/types/symbol.rs
