+++ src/writer.rs
@@ -1,9 +1,9 @@
 use crate::raw_symbol_token_ref::AsRawSymbolTokenRef;
 use crate::result::IonResult;
-use crate::types::decimal::Decimal;
-use crate::types::timestamp::Timestamp;
+use crate::types::Decimal;
+use crate::types::Int;
 use crate::types::IonType;
-use crate::Int;
+use crate::types::Timestamp;
 
 /**
  * This trait captures the format-agnostic encoding functionality needed to write native Rust types
